requirements.txt
Django>=5.2.5
Pillow>=10.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.7

# Testing and Quality Assurance
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
hypothesis>=6.82.0
factory-boy>=3.3.0
freezegun>=1.2.2

# Observability and Monitoring
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-django>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
prometheus-client>=0.17.1

# Resilience and Circuit Breaking
circuitbreaker>=1.4.0
django-ratelimit>=4.1.0

# Feature Flags and Configuration
django-environ>=0.11.2
django-waffle>=4.1.0

# Data Validation and Serialization
pydantic>=2.4.0
jsonschema>=4.19.0
